policies: fix dc aware and rack aware policies initialization
In cases when nodes are listed not in a proper groups of dc (for DCAwareRoundRobinPolicy) and dc+rack (for RackAwareRoundRobinPolicy).
Policy register only nodes from the last group or dc+rack.
So, policy knows only last group of nodes, rest of the nodes considered
to be non-existent, until node is restarted, which can be days.